lib/std: support riscv64
Borrow the ucontext layout from Go (probably reasonable?) and fix up
some missing imports for shared structures.

I can compile a glibc-linked hello.zig with:

   zig build-exe hello.zig -lc -target riscv64-linux-gnu.2.38 -mcpu=generic_rv64

or a native Zig binary with:

   zig build-exe hello.zig -target riscv64-linux -mcpu=generic_rv64

But the resulting binary segfaults immediately in my (perhaps broken) qemu
setup.  I've no experience in riscv64, so this maybe be broken in subtle
ways, but perhaps its a useful point for someone with the right hardware
to make more progress.

Builds on #18803